Refactor according to review requirements
